pyproject.toml
[tool.poetry]
name = "todo-list-django"
version = "0.1.0"
description = "Django TODO application"
authors = ["Igor Pejic <igor.pejic@dr.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg = "^3.1.13"
django-cities = "^0.6.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']  # You can specify the Python versions you're targeting.
include = '\.pyi?$'
